Ajustando metodo

+++ src/users/users.controller.ts
@@ -1,4 +1,12 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Res } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Res,
+} from '@nestjs/common';
 import { UsersService } from './users.service';
 import * as express from 'express';
 import { UserDomain } from './user.domain';
@@ -13,13 +21,18 @@ export class UsersController {
   }
 
   @Post()
+  // eslint-disable-next-line prettier/prettier
   async createUser(@Res() response: express.Response, @Body() user: UserDomain) {
     const userCreated = await this.usersService.createUser(user);
     return response.status(201).json(userCreated);
   }
 
-  @Delete()
-  deleteUser(@Res() response: express.Response, @Param() id: string) {
+  @Delete(':id')
+  deleteUser(@Res() response: express.Response, @Param('id') id: string) {
+    if (this.usersService.findUserById(id) === null) {
+      return response.status(404);
+    }
+
     this.usersService.deleteUserById(id);
     return response.status(204);
   }
